+++ window_scene/session/container/src/window_event_channel.cpp
@@ -209,8 +209,8 @@ void WindowEventChannel::PrintPointerEvent(const std::shared_ptr<MMI::PointerEve
             }
             WLOGFD("pointerId:%{public}d,DownTime:%{public}" PRId64 ",IsPressed:%{public}d,"
                 "DisplayX:%{public}d,DisplayY:%{public}d,WindowX:%{public}d,WindowY:%{public}d,",
-                pointerId, item.GetDownTime(), item.IsPressed(), item.GetDisplayX(), item.GetDisplayY(),
-                item.GetWindowX(), item.GetWindowY());
+                pointerId, item.GetDownTime(), item.IsPressed(), static_cast<int32_t>(item.GetDisplayX()), static_cast<int32_t>(item.GetDisplayY()),
+                static_cast<int32_t>(item.GetWindowX()), static_cast<int32_t>(item.GetWindowY()));
         }
     } else {
         PrintInfoPointerEvent(event);
